Check for existing change-id line in commit msg
`gerrit-upload` now checks before it starts the upload transient (only
if gerrit-use-gitreview-interface is nil) that the latest commit
contains a Change-Id line in its commit message. If it doesn't, it
downloads the commit-msg hook and raises an error. The user than has to
call git-commit --amend manually (or git rebase --interactive) to add
the missing changeid lines via the commit-msg hook.

Change-Id: I99a32345dc6ab9864f3b56fc9b9015006c9a8321