Handle non-unique change_ids in gerrit-rest-topic-*
The gerrit-rest-topic* functions fail when the topic contains changes
that have the same change_id (the Ixxxx string), which happens for
topics that contain cherry-picks. cherry-picks have the same change_id
as the original gerrit change from which it was cherry-picked.  This
issue is fixed in this commit, by using the unique-change-id (see
`gerrit-rest--change-info-to-unique-changeid).

Change-Id: Ia73c6175af24770363913e61cfa5944f87cbe3b2