Add gerrit-download-transient
This adds an initial version of the gerrit-download transient, which
currently does the same as the `gerrit-upload` command, but it allows
you to select a branch in the query that is used for displaying the
changes. As a next step, support for downloading changes from other
repos, i.e., not the current one, will be added.

Change-Id: I91c74b4f1e44330d00a014b04b744b5e9c52d798