Add support for downloading changes for different projects
Now it is possible to download a change for a project, which is not the
current one, i.e., to download a change for project B you don't have to
be in the workspace of project B to download it.

A new key (k) was added to the gerrit-download-transient to download
changes for known projects.

Change-Id: Ie4d55bced47af0575fce3a4a01aea6e0345771e1