(WIP) Improve support for selecting and downloading changes
This change adds support for selecting and downloading changes that
match a custom gerrit query (e.g. "is:open").

A new interactive function called gerrit-download-and-open-overview' was
added which downloads the change and opens an magit-commit buffer using
magit-show-commit'. (It the downloaded change is at the top of a
relation chain, the other commits are currently not displayed by this
`gerrit-download-and-open-overview' function.

Closes: #19
Change-Id: Ie3ed04a8f212cb8c0a98745d1dca4f680f420642