Add clickable links to fields in the gerrit-dashboard
Similar to the links in the gerrit webfrontend, gerrit-dashboard now
supports clicking on some fields (repo, change-id, branch, topic,
assignee, owner) in the dashboard. For most of the fields a new query is
performed and displayed via a new function called `gerrit-query`. For
the change-id only the patch is displayed in a new buffer.

Change-Id: Ibc2c257fd69d372214f4679bc4fc341f6c441ee0