Remove gerrit-rest--escape-project and add new sync-v2 function
gerrit-rest--escape-project can be removed in favor of the
url-build-query-string function, which escapes all fields in the query
string. gerrit-rest--escape-project was also used to determine the
unique-changeid. Since the branch name might also contain chars that
need to be escaped, url-hexify-string is used on the whole
unique-changeid string instead of gerrit-rest--escape-project.

This change also fises some docstring issues reported by emacs-29.

Change-Id: I9c72023a54c9d4444d9b9dbc7903aa7df2e60067