Fix Invalid application/json in request in some REST requests
I've noticed that the requests that take a payload with a "notify"
dict don't work because the value of the "notify" entry needs to be
"NONE" not "None".

In the course of this change, the gerrit-rest-sync-v2 method was
improved, because the http status code wasn't checked.  Since some
requests don't return json on success, the status code needs to be
checked.

Change-Id: I4bf07adecae1522bea492e8c45b18136348f974b